@@ -12,7 +12,7 @@
     "cookie-parser": "~1.4.3",
     "debug": "~2.2.0",
     "dotenv": "^4.0.0",
-    "express": "~4.13.4",
+    "express": "^4.16.3",
     "express-handlebars": "^3.0.0",
     "hbs": "^4.0.1",
     "moment": "^2.15.1",
@@ -25,18 +25,17 @@
     "babel-core": "^6.17.0",
     "babel-loader": "^6.2.5",
     "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.16.0",
     "css-loader": "^0.25.0",
     "html-webpack-plugin": "^2.22.0",
     "immutability-helper": "^2.0.0",
     "node-sass": "^3.10.1",
-    "react": "^15.3.2",
+    "react": "^16.3.2",
     "react-datetime": "^2.6.0",
-    "react-dom": "^15.3.2",
+    "react-dom": "^16.3.2",
     "react-form-data": "^0.2.3",
     "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",    
+    "style-loader": "^0.13.1",
     "webpack": "^1.13.2",
     "webpack-dev-server": "^1.16.2"
   },
